Update tests
Merge tests and test for 0 players
